Refactor KaldbConfig.
