+++ kaldb/src/main/java/com/slack/kaldb/metadata/core/KaldbMetadataStore.java
@@ -2,7 +2,7 @@ package com.slack.kaldb.metadata.core;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 import static com.google.common.base.Preconditions.checkState;
-import static com.slack.kaldb.config.KaldbConfig.DEFAULT_ZK_TIMEOUT_SECS;
+import static com.slack.kaldb.server.KaldbConfig.DEFAULT_ZK_TIMEOUT_SECS;
 
 import com.google.common.base.Function;
 import com.google.common.util.concurrent.Futures;
