+++ kaldb/src/main/java/com/slack/kaldb/metadata/core/UpdatableCacheableMetadataStore.java
@@ -3,9 +3,9 @@ package com.slack.kaldb.metadata.core;
 import com.google.common.util.concurrent.Futures;
 import com.google.common.util.concurrent.ListenableFuture;
 import com.google.protobuf.InvalidProtocolBufferException;
-import com.slack.kaldb.config.KaldbConfig;
 import com.slack.kaldb.metadata.zookeeper.InternalMetadataStoreException;
 import com.slack.kaldb.metadata.zookeeper.MetadataStore;
+import com.slack.kaldb.server.KaldbConfig;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
