+++ kaldb/src/test/java/com/slack/kaldb/elasticsearchApi/ElasticsearchApiServiceTest.java
@@ -1,6 +1,6 @@
 package com.slack.kaldb.elasticsearchApi;
 
-import static com.slack.kaldb.config.KaldbConfig.DEFAULT_START_STOP_DURATION;
+import static com.slack.kaldb.server.KaldbConfig.DEFAULT_START_STOP_DURATION;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 import brave.Tracing;
