+++ kaldb/src/test/java/com/slack/kaldb/recovery/RecoveryServiceTest.java
@@ -1,6 +1,6 @@
 package com.slack.kaldb.recovery;
 
-import static com.slack.kaldb.config.KaldbConfig.DEFAULT_START_STOP_DURATION;
+import static com.slack.kaldb.server.KaldbConfig.DEFAULT_START_STOP_DURATION;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.awaitility.Awaitility.await;
